+++ inst/themes/dark-mushroom.js
@@ -60,7 +60,7 @@
             splitLine: {
                 lineStyle: {
                     type: 'dashed',
-                    color: '#aaa'
+                    color: '#555'
                 }
             },
             splitArea: {
@@ -71,9 +71,8 @@
         };
     };
 
-    var colorPalette = [
-    	//'#cc0e00','#ff1a0a','#ff8880','#ffc180','#ffc2b0','#ffffff','#ff8880','#ffe6e6'  //old
-        '#d73027','#f46d43','#fdae61','#fee08b','#d9ef8b','#a6d96a','#66bd63','#1a9850'
+    var colorPalette = [   // carto pastel
+      "#66C5CC", "#F6CF71", "#F89C74", "#DCB0F2", "#87C55F", "#9EB9F3", "#FE88B1", "#C9DB74", "#8BE0A4", "#B497E7", "#D3B484"
     ];
     var theme = {
         color: colorPalette,