+++ tests/testthat/test-presets.R
@@ -144,7 +144,7 @@ test_that("presets with series.param", {
   expect_equal(p$x$opts$series[[1]]$stack, 'stk')
   p <- ec.init(ctype='line', 
      series.param= list(areaStyle= list(show= T), stack= 'stk', 
-                        data=list(c(0,0), c(2,2)))
+                        data= list(c(0,0), c(2,2)))
   )
   expect_equal(p$x$opts$series[[1]]$data[[2]], c(2,2))
 })